Add channel picker themes. Consolidate commonly used css functions.
