+++ packages/mgt/src/styles/theme.scss
@@ -1,8 +1,21 @@
 @import '../components/mgt-people-picker/mgt-people-picker.theme.scss';
+@import '../components/mgt-teams-channel-picker/mgt-teams-channel-picker.theme.scss';
+
+$themes: ('light', 'dark');
+$theme-default: 'light';
+
+@function set-var($var, $theme, $component) {
+  $theme-value: map-get($component, $var);
+
+  @return var(#{map-get($theme-value, '_var')}, #{map-get($theme-value, $theme)});
+}
 
 @each $theme in $themes {
   .mgt-#{$theme} {
-    @each $name, $var in $theme-values {
+    @each $name, $var in $people-picker {
+      #{map-get($var, _var)}: #{map-get($var, $theme)};
+    }
+    @each $name, $var in $channel-picker {
       #{map-get($var, _var)}: #{map-get($var, $theme)};
     }
   }
